[Obs AI] Remove Obs Agent (#262937)
Closes https://github.com/elastic/obs-ai-team/issues/562

## Summary

This PR removes all remaining **Observability Agent** code as a
follow-up to the migration to **Elastic Agent and skills**.

It also updates references to `OBSERVABILITY_AGENT_ID` to use the
**Elastic Agent ID** where required, ensuring consistency with the new
architecture.

### What’s changed

-  Removed all **Observability Agent** related code
- Replaced usages of `OBSERVABILITY_AGENT_ID` with the **Elastic Agent
ID**
- Removed logic that made the Observability Agent the default experience

### Context

- Migration work was introduced in:
  https://github.com/elastic/obs-ai-team/issues/561

- Evaluations were already updated in:
  https://github.com/elastic/kibana/pull/262293

This PR completes the cleanup phase after the migration.

---------

(cherry picked from commit 76b9c017e2ab131ebdeeddee9864dd7967c1348f)
