+++ src/platform/packages/shared/kbn-synthtrace/src/scenarios/agent_builder/tools/get_logs/missing_db_index_pool_exhaustion.ts
@@ -59,12 +59,11 @@
  * }
  * ```
  *
- * Validate via Observability agent (with log search skill):
+ * Validate via Elastic AI Agent (with log search skill):
  *
  * ```
  * POST kbn:///api/agent_builder/converse
  * {
- *   "agent_id": "observability.agent",
  *   "input": "Something is wrong with checkout. Can you investigate the logs?"
  * }
  * ```
