+++ x-pack/solutions/observability/plugins/apm/public/components/app/error_group_details/index.tsx
@@ -11,10 +11,7 @@ import React, { useEffect } from 'react';
 import { omit } from 'lodash';
 import { useHistory } from 'react-router-dom';
 import { usePerformanceContext } from '@kbn/ebt-tools';
-import {
-  OBSERVABILITY_AGENT_ID,
-  OBSERVABILITY_ERROR_ATTACHMENT_TYPE_ID,
-} from '@kbn/observability-agent-builder-plugin/public';
+import { OBSERVABILITY_ERROR_ATTACHMENT_TYPE_ID } from '@kbn/observability-agent-builder-plugin/public';
 import { isOpenTelemetryAgentName, isRumAgentName } from '../../../../common/agent_name';
 import { NOT_AVAILABLE_LABEL } from '../../../../common/i18n';
 import { useApmServiceContext } from '../../../context/apm_service/use_apm_service_context';
@@ -210,7 +207,6 @@ export function ErrorGroupDetails() {
 
     agentBuilder.setChatConfig({
       newConversation: true,
-      agentId: OBSERVABILITY_AGENT_ID,
       attachments: [
         {
           type: OBSERVABILITY_ERROR_ATTACHMENT_TYPE_ID,
