+++ x-pack/solutions/observability/plugins/apm/public/components/app/transaction_details/index.test.tsx
@@ -254,7 +254,6 @@ describe('TransactionDetails', () => {
     );
 
     expect(mockSetAgentBuilderChatConfig).toHaveBeenCalledWith({
-      agentId: 'observability.agent',
       attachments: [
         expect.objectContaining({
           type: 'observability.transaction',
