+++ x-pack/solutions/observability/plugins/apm/public/components/app/transaction_details/index.tsx
@@ -7,10 +7,7 @@
 
 import { EuiSpacer, EuiTitle } from '@elastic/eui';
 import { i18n } from '@kbn/i18n';
-import {
-  OBSERVABILITY_AGENT_ID,
-  OBSERVABILITY_TRANSACTION_ATTACHMENT_TYPE_ID,
-} from '@kbn/observability-agent-builder-plugin/public';
+import { OBSERVABILITY_TRANSACTION_ATTACHMENT_TYPE_ID } from '@kbn/observability-agent-builder-plugin/public';
 import React, { useEffect } from 'react';
 import { Redirect, useHistory } from 'react-router-dom';
 import { useApmServiceContext } from '../../../context/apm_service/use_apm_service_context';
@@ -78,7 +75,6 @@ export function TransactionDetails() {
     }
 
     agentBuilder.setChatConfig({
-      agentId: OBSERVABILITY_AGENT_ID,
       attachments: [
         {
           type: OBSERVABILITY_TRANSACTION_ATTACHMENT_TYPE_ID,
