+++ x-pack/solutions/observability/plugins/infra/public/components/asset_details/hooks/use_host_attachment_config.test.ts
@@ -126,7 +126,6 @@ describe('useHostAttachmentConfig', () => {
     renderHook(() => useHostAttachmentConfig());
 
     expect(mockSetAgentBuilderChatConfig).toHaveBeenCalledWith({
-      agentId: 'observability.agent',
       attachments: [
         expect.objectContaining({
           type: 'observability.host',
