+++ x-pack/solutions/observability/plugins/infra/public/components/asset_details/hooks/use_host_attachment_config.ts
@@ -7,10 +7,7 @@
 
 import { useEffect } from 'react';
 import { i18n } from '@kbn/i18n';
-import {
-  OBSERVABILITY_AGENT_ID,
-  OBSERVABILITY_HOST_ATTACHMENT_TYPE_ID,
-} from '@kbn/observability-agent-builder-plugin/public';
+import { OBSERVABILITY_HOST_ATTACHMENT_TYPE_ID } from '@kbn/observability-agent-builder-plugin/public';
 import { useKibanaContextForPlugin } from '../../../hooks/use_kibana';
 import { useAssetDetailsRenderPropsContext } from './use_asset_details_render_props';
 import { useDatePickerContext } from './use_date_picker';
@@ -33,7 +30,6 @@ export const useHostAttachmentConfig = () => {
     }
 
     agentBuilder.setChatConfig({
-      agentId: OBSERVABILITY_AGENT_ID,
       attachments: [
         {
           type: OBSERVABILITY_HOST_ATTACHMENT_TYPE_ID,
