+++ x-pack/solutions/observability/plugins/observability/public/pages/alert_details/alert_details.tsx
@@ -39,10 +39,7 @@ import { css } from '@emotion/react';
 import { omit } from 'lodash';
 import { usePageReady } from '@kbn/ebt-tools';
 import moment from 'moment';
-import {
-  OBSERVABILITY_AGENT_ID,
-  OBSERVABILITY_ALERT_ATTACHMENT_TYPE_ID,
-} from '@kbn/observability-agent-builder-plugin/public';
+import { OBSERVABILITY_ALERT_ATTACHMENT_TYPE_ID } from '@kbn/observability-agent-builder-plugin/public';
 import { ObsCasesContext } from './components/obs_cases_context';
 import { RelatedAlerts } from './components/related_alerts/related_alerts';
 import type { AlertDetailsSource, TabId } from './types';
@@ -193,7 +190,6 @@ export function AlertDetails() {
 
     agentBuilder.setChatConfig({
       newConversation: true,
-      agentId: OBSERVABILITY_AGENT_ID,
       attachments: [
         {
           id: alertUuid,
