+++ x-pack/solutions/observability/plugins/observability_agent_builder/common/constants.ts
@@ -5,9 +5,6 @@
  * 2.0.
  */
 
-// Agent IDs
-export const OBSERVABILITY_AGENT_ID = 'observability.agent';
-
 // Inference Feature Registry IDs
 export const OBSERVABILITY_AI_INSIGHTS_INFERENCE_PARENT_FEATURE_ID =
   'observability_ai_insights_inference_parent_feature';
