+++ x-pack/solutions/observability/plugins/observability_agent_builder/common/index.ts
@@ -6,7 +6,6 @@
  */
 
 export {
-  OBSERVABILITY_AGENT_ID,
   OBSERVABILITY_AI_INSIGHT_ATTACHMENT_TYPE_ID,
   OBSERVABILITY_ALERT_ATTACHMENT_TYPE_ID,
   OBSERVABILITY_ERROR_ATTACHMENT_TYPE_ID,
