+++ x-pack/solutions/observability/plugins/observability_agent_builder/public/components/ai_insight/ai_insight.test.tsx
@@ -15,7 +15,6 @@ import { useKibana } from '../../hooks/use_kibana';
 import { useLicense } from '../../hooks/use_license';
 import { useGenAIConnectors } from '../../hooks/use_genai_connectors';
 import { useStreamingAiInsight } from '../../hooks/use_streaming_ai_insight';
-import { OBSERVABILITY_AGENT_ID } from '../../../common/constants';
 
 jest.mock('@kbn/kibana-react-plugin/public', () => ({
   useUiSetting$: jest.fn(),
@@ -222,7 +221,6 @@ describe('AiInsight', () => {
       expect(mockOpenChat).toHaveBeenCalledWith({
         newConversation: true,
         attachments: [{ type: 'test', data: {} }],
-        agentId: OBSERVABILITY_AGENT_ID,
       });
 
       unmount();
