+++ x-pack/solutions/observability/plugins/observability_agent_builder/public/index.ts
@@ -9,7 +9,6 @@ import type { PluginInitializerContext } from '@kbn/core/public';
 import { ObservabilityAgentBuilderPlugin } from './plugin';
 
 export {
-  OBSERVABILITY_AGENT_ID,
   OBSERVABILITY_AI_INSIGHT_ATTACHMENT_TYPE_ID,
   OBSERVABILITY_ALERT_ATTACHMENT_TYPE_ID,
   OBSERVABILITY_ERROR_ATTACHMENT_TYPE_ID,
