--- x-pack/solutions/observability/plugins/observability_agent_builder/server/agent/register_observability_agent.ts
@@ -1,164 +1,0 @@
-/*
- * Copyright Elasticsearch B.V. and/or licensed to Elasticsearch B.V. under one
- * or more contributor license agreements. Licensed under the Elastic License
- * 2.0; you may not use this file except in compliance with the Elastic License
- * 2.0.
- */
-
-import type { Logger } from '@kbn/core/server';
-import dedent from 'dedent';
-import type {
-  ObservabilityAgentBuilderCoreSetup,
-  ObservabilityAgentBuilderPluginSetupDependencies,
-} from '../types';
-import { OBSERVABILITY_GET_INDEX_INFO_TOOL_ID } from '../tools';
-import { getAgentBuilderResourceAvailability } from '../utils/get_agent_builder_resource_availability';
-import { OBSERVABILITY_AGENT_ID } from '../../common/constants';
-import { OBSERVABILITY_TOOL_IDS, PLATFORM_TOOL_IDS } from '../tools/register_tools';
-
-const OBSERVABILITY_AGENT_TOOL_IDS = [...PLATFORM_TOOL_IDS, ...OBSERVABILITY_TOOL_IDS];
-
-export async function registerObservabilityAgent({
-  core,
-  plugins,
-  logger,
-}: {
-  core: ObservabilityAgentBuilderCoreSetup;
-  plugins: ObservabilityAgentBuilderPluginSetupDependencies;
-  logger: Logger;
-}) {
-  plugins.agentBuilder?.agents.register({
-    id: OBSERVABILITY_AGENT_ID,
-    name: 'Observability Agent',
-    description: 'Agent specialized in logs, metrics, and traces',
-    avatar_icon: 'logoObservability',
-    availability: {
-      cacheMode: 'space',
-      handler: async ({ request }) => {
-        return getAgentBuilderResourceAvailability({ core, request, logger });
-      },
-    },
-    configuration: ({ request }) => {
-      const urlPrefix = core.http.basePath.get(request);
-
-      return {
-        instructions:
-          dedent(`You are an observability specialist agent that helps Site Reliability Engineers (SREs) investigate incidents and understand system health.
-
-        ${getTimeRangeInstructions()}
-        ${getInvestigationInstructions()}
-        ${getReasoningInstructions()}
-        ${getTraceMetricFormatInstructions()}
-        ${getFieldDiscoveryInstructions()}
-        ${getKqlInstructions()}
-        ${getEntityLinkingInstructions({ urlPrefix })}
-      `),
-        tools: [{ tool_ids: OBSERVABILITY_AGENT_TOOL_IDS }],
-      };
-    },
-  });
-
-  logger.debug('Successfully registered observability agent in agent-builder');
-}
-
-function getTimeRangeInstructions() {
-  return dedent(`
-    <time_range>
-    ### Time Range
-    The screen context attachment contains the user's currently selected time range (e.g. "Time range: now-15m to now").
-    When a time range is available from screen context or the user's explicit request, always pass it as the \`start\` and \`end\` parameters to tool calls — do not rely on tool defaults when explicit time context is available.
-    </time_range>
-  `);
-}
-
-function getInvestigationInstructions() {
-  return dedent(`
-    <investigation_approach>
-    ### Investigation Approach
-    Follow a progressive workflow - start broad, then narrow down:
-    1. **Triage**: What's the severity? How many users/services affected?
-    2. **Scope**: Which components are affected? What's the blast radius?
-    3. **Timeline**: When did it start? What changed before symptoms appeared?
-    4. **Correlation**: What error patterns exist? What's the sequence of events?
-    5. **Root Cause**: Distinguish the SOURCE (where the problem started) from AFFECTED services (impacted downstream)
-    6. **Verification**: Does your hypothesis explain ALL the symptoms? If not, dig deeper.
-    </investigation_approach>
-  `);
-}
-
-function getReasoningInstructions() {
-  return dedent(`
-    <reasoning_principles>
-    ### Reasoning Principles
-    - **Be quantitative**: Quote specific metrics (error rate %, latency ms, throughput rpm). Avoid vague terms like "high" without numbers.
-    - **Correlation ≠ causation**: Look for temporal sequence (what happened FIRST) and causal mechanism.
-    - **Consider all layers**: Infrastructure (CPU, memory, disk) → Application (latency, throughput, failure rate) → Dependencies (databases, caches, external APIs).
-    - **Follow evidence**: Support hypotheses with data. Acknowledge uncertainty when evidence is inconclusive.
-    </reasoning_principles>
-  `);
-}
-
-function getFieldDiscoveryInstructions() {
-  return dedent(`
-    <field_discovery>
-    ### Field Discovery
-    Before using field names in \`groupBy\`, \`kqlFilter\`, or \`aggregation.field\` parameters, call \`${OBSERVABILITY_GET_INDEX_INFO_TOOL_ID}\` first.
-    Clusters use different naming conventions (ECS vs OpenTelemetry) - discovering fields first prevents errors.
-    </field_discovery>
-  `);
-}
-
-export function getKqlInstructions() {
-  return dedent(`
-    <kql_syntax>
-    ### KQL (Kibana Query Language)
-    Use KQL syntax for \`kqlFilter\` parameters:
-    - Match: \`field: value\`, \`field: (a OR b OR c)\`
-    - Range: \`field > 100\`, \`field >= 10 AND field <= 20\`
-    - Wildcards: \`field: prefix*\` (trailing only)
-    - Negation: \`NOT field: value\`, \`NOT message: "noisy string"\`
-    - Logical operators: Combine with \`AND\`/\`OR\`, \`(field: value OR field: value) AND field: value\`, use parentheses for precedence
-    - Exists: \`field: *\` (field has any value)
-    - Phrases: \`message: "connection refused"\` (exact phrase in text)
-    </kql_syntax>
-  `);
-}
-
-function getTraceMetricFormatInstructions() {
-  return dedent(`
-    ### TRACE METRIC FORMATS
-    All observability tools use these standardized units for trace metrics:
-    - **Latency**: milliseconds (ms)
-    - **Throughput**: transactions per minute (tpm)
-    - **Failure rate**: 0-1 scale (e.g., 0.05 = 5% failure rate)
-  `);
-}
-
-/**
- * Entity Linking instructions for the Observability Agent.
- * Instructs the LLM to format entities as clickable links using Kibana's relative URL paths.
- */
-export function getEntityLinkingInstructions({ urlPrefix }: { urlPrefix: string }): string {
-  return dedent(`
-  <entity_linking>
-  ### Entity Linking Guidelines
-  Use markdown for readability. When referencing entities, create clickable links.
-  **IMPORTANT**: Do NOT wrap links in backticks - backticks prevent links from being clickable.
-
-  | Entity | Link Format | Example |
-  |--------|-------------|---------|
-  | Service | [<serviceName>](${urlPrefix}/app/apm/services/<serviceName>) | "The [payments](${urlPrefix}/app/apm/services/payments) service is experiencing high latency." |
-  | Transaction | [<transactionName>](${urlPrefix}/app/apm/services/<serviceName>/transactions) | "The transaction [POST /checkout](${urlPrefix}/app/apm/services/payments/transactions) took 500ms." |
-  | Trace | [<traceId>](${urlPrefix}/app/apm/link-to/trace/<traceId>) | "See trace [8bc26008603e16819bd6fcfb80fceff5](${urlPrefix}/app/apm/link-to/trace/8bc26008603e16819bd6fcfb80fceff5)" |
-  | Error | [<errorKey>](${urlPrefix}/app/apm/services/<serviceName>/errors/<errorKey>) | "Error [upstream-5xx](${urlPrefix}/app/apm/services/catalog-api/errors/upstream-5xx) suggests a dependency failure." |
-  | Service Errors | [errors](${urlPrefix}/app/apm/services/<serviceName>/errors) | "Review all [errors](${urlPrefix}/app/apm/services/frontend/errors) for the [frontend](${urlPrefix}/app/apm/services/frontend) service." |
-  | Service Logs | [logs](${urlPrefix}/app/apm/services/<serviceName>/logs) | "Check [logs](${urlPrefix}/app/apm/services/frontend/logs) for the [frontend](${urlPrefix}/app/apm/services/frontend) service." |
-  | Host | [<hostName>](${urlPrefix}/app/metrics/detail/host/<hostName>) | "Host [web-01](${urlPrefix}/app/metrics/detail/host/web-01) is experiencing high CPU usage." |
-  | Service Map | [Service Map](${urlPrefix}/app/apm/services/<serviceName>/service-map) | "Check the [Service Map](${urlPrefix}/app/apm/services/payments/service-map) to see dependencies." |
-  | Dependencies | [Dependencies](${urlPrefix}/app/apm/services/<serviceName>/dependencies) | "View [Dependencies](${urlPrefix}/app/apm/services/catalog-api/dependencies) to identify upstream issues." |
-  | Alert | [<alertId>](${urlPrefix}/app/observability/alerts/<alertId>) | "Alert [alert-uuid-123](${urlPrefix}/app/observability/alerts/alert-uuid-123) was triggered." |
-  | Alert Rules | [<alertRuleId>](${urlPrefix}/app/rules/rule/<alertRuleId>) | "Alert Rule [alert-uuid-123](${urlPrefix}/app/rules/rule/alert-uuid-123)." |
-  | Discover | [Discover](${urlPrefix}/app/discover) | "Go to [Discover](${urlPrefix}/app/discover) to investigate the issue further." |
-  </entity_linking>
-`);
-}
