+++ x-pack/solutions/observability/plugins/observability_agent_builder/server/routes/ai_insights/alert_ai_insights/generate_alert_ai_insight.ts
@@ -21,7 +21,7 @@ import type {
   ObservabilityAgentBuilderCoreSetup,
   ObservabilityAgentBuilderPluginSetupDependencies,
 } from '../../../types';
-import { getEntityLinkingInstructions } from '../../../agent/register_observability_agent';
+import { getEntityLinkingInstructions } from '../../../utils/get_entity_linking_instructions';
 import { runSignalFetchers, formatSignalResults } from './signal_fetchers';
 
 /**
