+++ x-pack/solutions/observability/plugins/observability_shared/moon.yml
@@ -32,7 +32,6 @@ dependsOn:
   - '@kbn/es-types'
   - '@kbn/spaces-plugin'
   - '@kbn/core-application-browser'
-  - '@kbn/core-application-common'
   - '@kbn/datemath'
   - '@kbn/rison'
   - '@kbn/kibana-utils-plugin'
