+++ x-pack/solutions/observability/plugins/slo/public/embeddable/slo/common/slo_overview_details.tsx
@@ -24,10 +24,7 @@ import type { SLOWithSummaryResponse } from '@kbn/slo-schema';
 import React, { useCallback, useEffect, useState } from 'react';
 import type { SloTabId } from '@kbn/deeplinks-observability';
 import { OVERVIEW_TAB_ID } from '@kbn/deeplinks-observability';
-import {
-  OBSERVABILITY_AGENT_ID,
-  OBSERVABILITY_SLO_ATTACHMENT_TYPE_ID,
-} from '@kbn/observability-agent-builder-plugin/public';
+import { OBSERVABILITY_SLO_ATTACHMENT_TYPE_ID } from '@kbn/observability-agent-builder-plugin/public';
 import { HeaderTitle } from '../../../pages/slo_details/components/header_title';
 import { SloDetails } from '../../../pages/slo_details/components/slo_details';
 import { useSloDetailsTabs } from '../../../pages/slo_details/hooks/use_slo_details_tabs';
@@ -72,7 +69,6 @@ export function SloOverviewDetailsContent({
     }
 
     agentBuilder.setChatConfig({
-      agentId: OBSERVABILITY_AGENT_ID,
       attachments: [
         {
           type: OBSERVABILITY_SLO_ATTACHMENT_TYPE_ID,
